@@ -2,22 +2,23 @@ package autopeering
 
 import (
 	"encoding/base64"
+	"errors"
 	"fmt"
 	"net"
 	"strings"
 
-	"github.com/iotaledger/goshimmer/packages/autopeering/discover"
-	"github.com/iotaledger/goshimmer/packages/autopeering/peer"
-	"github.com/iotaledger/goshimmer/packages/autopeering/peer/service"
-	"github.com/iotaledger/goshimmer/packages/autopeering/selection"
-	"github.com/iotaledger/goshimmer/packages/autopeering/server"
-	"github.com/iotaledger/goshimmer/packages/autopeering/transport"
-	"github.com/iotaledger/goshimmer/packages/parameter"
-	"github.com/iotaledger/goshimmer/plugins/autopeering/local"
-	"github.com/iotaledger/goshimmer/plugins/gossip"
+	"github.com/iotaledger/hive.go/autopeering/discover"
+	"github.com/iotaledger/hive.go/autopeering/peer"
+	"github.com/iotaledger/hive.go/autopeering/peer/service"
+	"github.com/iotaledger/hive.go/autopeering/selection"
+	"github.com/iotaledger/hive.go/autopeering/server"
+	"github.com/iotaledger/hive.go/autopeering/transport"
 	"github.com/iotaledger/hive.go/logger"
 	"github.com/iotaledger/hive.go/node"
-	"github.com/pkg/errors"
+
+	"github.com/iotaledger/goshimmer/plugins/autopeering/local"
+	"github.com/iotaledger/goshimmer/plugins/config"
+	"github.com/iotaledger/goshimmer/plugins/gossip"
 )
 
 var (
@@ -25,6 +26,8 @@ var (
 	Discovery *discover.Protocol
 	// Selection is the peer selection protocol.
 	Selection *selection.Protocol
+
+	ErrParsingMasterNode = errors.New("can't parse master node")
 
 	log *logger.Logger
 )
@@ -36,42 +39,64 @@ func configureAP() {
 	}
 	log.Debugf("Master peers: %v", masterPeers)
 
-	Discovery = discover.New(local.GetInstance(), discover.Config{
-		Log:         log.Named("disc"),
-		MasterPeers: masterPeers,
-	})
+	Discovery = discover.New(local.GetInstance(),
+		discover.Logger(log.Named("disc")),
+		discover.Version(config.Node.GetUint32(CFG_VERSION)),
+		discover.MasterPeers(masterPeers),
+	)
 
+	log.Infof("Protocol Version: %v", discover.VersionNum)
 	// enable peer selection only when gossip is enabled
 	if !node.IsSkipped(gossip.PLUGIN) {
-		Selection = selection.New(local.GetInstance(), Discovery, selection.Config{
-			Log:              log.Named("sel"),
-			RequiredServices: []service.Key{service.GossipKey},
-		})
+		Selection = selection.New(local.GetInstance(), Discovery,
+			selection.Logger(log.Named("sel")),
+			selection.NeighborValidator(selection.ValidatorFunc(isValidNeighbor)),
+		)
 	}
 }
 
+// isValidNeighbor checks whether a peer is a valid neighbor.
+func isValidNeighbor(p *peer.Peer) bool {
+	// gossip must be supported
+	gossipAddr := p.Services().Get(service.GossipKey)
+	if gossipAddr == nil {
+		return false
+	}
+	// the host for the gossip and peering service must be identical
+	gossipHost, _, err := net.SplitHostPort(gossipAddr.String())
+	if err != nil {
+		return false
+	}
+	peeringAddr := p.Services().Get(service.PeeringKey)
+	peeringHost, _, err := net.SplitHostPort(peeringAddr.String())
+	if err != nil {
+		return false
+	}
+	return gossipHost == peeringHost
+}
+
 func start(shutdownSignal <-chan struct{}) {
-	defer log.Info("Stopping Auto Peering server ... done")
+	defer log.Info("Stopping " + name + " ... done")
 
-	addr := local.GetInstance().Services().Get(service.PeeringKey)
-	udpAddr, err := net.ResolveUDPAddr(addr.Network(), addr.String())
+	lPeer := local.GetInstance()
+	// use the port of the peering service
+	peeringAddr := lPeer.Services().Get(service.PeeringKey)
+	_, peeringPort, err := net.SplitHostPort(peeringAddr.String())
 	if err != nil {
-		log.Fatalf("ResolveUDPAddr: %v", err)
+		panic(err)
 	}
-
-	// if the ip is an external ip, set it to unspecified
-	if udpAddr.IP.IsGlobalUnicast() {
-		if udpAddr.IP.To4() != nil {
-			udpAddr.IP = net.IPv4zero
-		} else {
-			udpAddr.IP = net.IPv6unspecified
-		}
-	}
-
-	conn, err := net.ListenUDP(addr.Network(), udpAddr)
+	// resolve the bind address
+	address := net.JoinHostPort(config.Node.GetString(local.CFG_BIND), peeringPort)
+	localAddr, err := net.ResolveUDPAddr(peeringAddr.Network(), address)
 	if err != nil {
-		log.Fatalf("ListenUDP: %v", err)
+		log.Fatalf("Error resolving %s: %v", local.CFG_BIND, err)
 	}
+
+	conn, err := net.ListenUDP(peeringAddr.Network(), localAddr)
+	if err != nil {
+		log.Fatalf("Error listening: %v", err)
+	}
+	defer conn.Close()
 
 	// use the UDP connection for transport
 	trans := transport.Conn(conn, func(network, address string) (net.Addr, error) { return net.ResolveUDPAddr(network, address) })
@@ -83,7 +108,7 @@ func start(shutdownSignal <-chan struct{}) {
 	}
 
 	// start a server doing discovery and peering
-	srv := server.Listen(local.GetInstance(), trans, log.Named("srv"), handlers...)
+	srv := server.Serve(lPeer, trans, log.Named("srv"), handlers...)
 	defer srv.Close()
 
 	// start the discovery on that connection
@@ -96,26 +121,29 @@ func start(shutdownSignal <-chan struct{}) {
 		defer Selection.Close()
 	}
 
-	log.Infof("Auto Peering started: address=%s", srv.LocalAddr())
-	log.Debugf("Auto Peering server started: PubKey=%s", base64.StdEncoding.EncodeToString(local.GetInstance().PublicKey()))
+	log.Infof("%s started: ID=%s Address=%s/%s", name, lPeer.ID(), peeringAddr.String(), peeringAddr.Network())
 
 	<-shutdownSignal
-	log.Info("Stopping Auto Peering server ...")
+
+	log.Infof("Stopping %s ...", name)
+
+	count := lPeer.Database().PersistSeeds()
+	log.Infof("%d peers persisted as seeds", count)
 }
 
 func parseEntryNodes() (result []*peer.Peer, err error) {
-	for _, entryNodeDefinition := range parameter.NodeConfig.GetStringSlice(CFG_ENTRY_NODES) {
+	for _, entryNodeDefinition := range config.Node.GetStringSlice(CFG_ENTRY_NODES) {
 		if entryNodeDefinition == "" {
 			continue
 		}
 
 		parts := strings.Split(entryNodeDefinition, "@")
 		if len(parts) != 2 {
-			return nil, fmt.Errorf("parseMaster")
+			return nil, fmt.Errorf("%w: master node parts must be 2, is %d", ErrParsingMasterNode, len(parts))
 		}
 		pubKey, err := base64.StdEncoding.DecodeString(parts[0])
 		if err != nil {
-			return nil, errors.Wrap(err, "parseMaster")
+			return nil, fmt.Errorf("%w: can't decode public key: %s", ErrParsingMasterNode, err)
 		}
 
 		services := service.New()
