@@ -1,8 +1,8 @@
 package gossip
 
 import (
-	"github.com/iotaledger/goshimmer/packages/autopeering/peer"
-	"github.com/iotaledger/goshimmer/packages/autopeering/peer/service"
+	"github.com/iotaledger/hive.go/autopeering/peer"
+	"github.com/iotaledger/hive.go/autopeering/peer/service"
 )
 
 // IsSupported returns whether the peer supports the gossip service.
