@@ -7,10 +7,12 @@ import (
 	"time"
 
 	"github.com/golang/protobuf/proto"
-	"github.com/iotaledger/goshimmer/packages/autopeering/peer"
-	"github.com/iotaledger/goshimmer/packages/autopeering/peer/service"
+	"github.com/iotaledger/goshimmer/packages/binary/tangle/model/transaction"
 	pb "github.com/iotaledger/goshimmer/packages/gossip/proto"
 	"github.com/iotaledger/goshimmer/packages/gossip/server"
+	"github.com/iotaledger/hive.go/autopeering/peer"
+	"github.com/iotaledger/hive.go/autopeering/peer/service"
+	"github.com/iotaledger/hive.go/database/mapdb"
 	"github.com/iotaledger/hive.go/events"
 	"github.com/iotaledger/hive.go/logger"
 	"github.com/stretchr/testify/assert"
@@ -25,47 +27,7 @@ var (
 	testTxData = []byte("testTx")
 )
 
-func getTestTransaction([]byte) ([]byte, error) { return testTxData, nil }
-
-func getTCPAddress(t require.TestingT) string {
-	tcpAddr, err := net.ResolveTCPAddr("tcp", "localhost:0")
-	require.NoError(t, err)
-	lis, err := net.ListenTCP("tcp", tcpAddr)
-	require.NoError(t, err)
-
-	addr := lis.Addr().String()
-	require.NoError(t, lis.Close())
-
-	return addr
-}
-
-func newTestManager(t require.TestingT, name string) (*Manager, func(), *peer.Peer) {
-	l := log.Named(name)
-	db := peer.NewMemoryDB(l.Named("db"))
-	local, err := peer.NewLocal("peering", name, db)
-	require.NoError(t, err)
-
-	// enable TCP gossipping
-	require.NoError(t, local.UpdateService(service.GossipKey, "tcp", getTCPAddress(t)))
-
-	mgr := NewManager(local, getTestTransaction, l)
-
-	srv, err := server.ListenTCP(local, l)
-	require.NoError(t, err)
-
-	// update the service with the actual address
-	require.NoError(t, local.UpdateService(service.GossipKey, srv.LocalAddr().Network(), srv.LocalAddr().String()))
-
-	// start the actual gossipping
-	mgr.Start(srv)
-
-	detach := func() {
-		mgr.Close()
-		srv.Close()
-		db.Close()
-	}
-	return mgr, detach, &local.Peer
-}
+func getTestTransaction(transaction.Id) ([]byte, error) { return testTxData, nil }
 
 func TestClose(t *testing.T) {
 	_, detach := newEventMock(t)
@@ -349,7 +311,7 @@ func TestDropUnsuccessfulAccept(t *testing.T) {
 	_, closeB, peerB := newTestManager(t, "B")
 	defer closeB()
 
-	e.On("connectionFailed", peerB).Once()
+	e.On("connectionFailed", peerB, mock.Anything).Once()
 
 	err := mgrA.AddInbound(peerB)
 	assert.Error(t, err)
@@ -409,6 +371,42 @@ func TestTxRequest(t *testing.T) {
 	e.AssertExpectations(t)
 }
 
+func newTestDB(t require.TestingT) *peer.DB {
+	db, err := peer.NewDB(mapdb.NewMapDB())
+	require.NoError(t, err)
+	return db
+}
+
+func newTestManager(t require.TestingT, name string) (*Manager, func(), *peer.Peer) {
+	l := log.Named(name)
+
+	services := service.New()
+	services.Update(service.PeeringKey, "peering", name)
+	local, err := peer.NewLocal(services, newTestDB(t))
+	require.NoError(t, err)
+
+	laddr, err := net.ResolveTCPAddr("tcp", "127.0.0.1:0")
+	require.NoError(t, err)
+	lis, err := net.ListenTCP("tcp", laddr)
+	require.NoError(t, err)
+
+	// enable TCP gossipping
+	require.NoError(t, local.UpdateService(service.GossipKey, lis.Addr().Network(), lis.Addr().String()))
+
+	srv := server.ServeTCP(local, lis, l)
+
+	// start the actual gossipping
+	mgr := NewManager(local, getTestTransaction, l)
+	mgr.Start(srv)
+
+	detach := func() {
+		mgr.Close()
+		srv.Close()
+		_ = lis.Close()
+	}
+	return mgr, detach, &local.Peer
+}
+
 func newEventMock(t mock.TestingT) (*eventMock, func()) {
 	e := &eventMock{}
 	e.Test(t)
@@ -435,7 +433,7 @@ type eventMock struct {
 	mock.Mock
 }
 
-func (e *eventMock) connectionFailed(p *peer.Peer)                    { e.Called(p) }
+func (e *eventMock) connectionFailed(p *peer.Peer, err error)         { e.Called(p, err) }
 func (e *eventMock) neighborAdded(n *Neighbor)                        { e.Called(n) }
 func (e *eventMock) neighborRemoved(p *peer.Peer)                     { e.Called(p) }
 func (e *eventMock) transactionReceived(ev *TransactionReceivedEvent) { e.Called(ev) }
