@@ -16,6 +16,7 @@ var lock sync.Mutex
 
 func Configure(plugin *node.Plugin) {
 	server.Events.AddNode.Attach(events.NewClosure(func(nodeId string) {
+		plugin.Node.Logger.Debugw("AddNode", "nodeID", nodeId)
 		lock.Lock()
 		defer lock.Unlock()
 
@@ -25,6 +26,7 @@ func Configure(plugin *node.Plugin) {
 	}))
 
 	server.Events.RemoveNode.Attach(events.NewClosure(func(nodeId string) {
+		plugin.Node.Logger.Debugw("RemoveNode", "nodeID", nodeId)
 		lock.Lock()
 		defer lock.Unlock()
 
@@ -32,6 +34,7 @@ func Configure(plugin *node.Plugin) {
 	}))
 
 	server.Events.NodeOnline.Attach(events.NewClosure(func(nodeId string) {
+		plugin.Node.Logger.Debugw("NodeOnline", "nodeID", nodeId)
 		lock.Lock()
 		defer lock.Unlock()
 
@@ -39,6 +42,7 @@ func Configure(plugin *node.Plugin) {
 	}))
 
 	server.Events.NodeOffline.Attach(events.NewClosure(func(nodeId string) {
+		plugin.Node.Logger.Debugw("NodeOffline", "nodeID", nodeId)
 		lock.Lock()
 		defer lock.Unlock()
 
@@ -46,6 +50,7 @@ func Configure(plugin *node.Plugin) {
 	}))
 
 	server.Events.ConnectNodes.Attach(events.NewClosure(func(sourceId string, targetId string) {
+		plugin.Node.Logger.Debugw("ConnectNodes", "sourceID", sourceId, "targetId", targetId)
 		lock.Lock()
 		defer lock.Unlock()
 
@@ -59,6 +64,7 @@ func Configure(plugin *node.Plugin) {
 	}))
 
 	server.Events.DisconnectNodes.Attach(events.NewClosure(func(sourceId string, targetId string) {
+		plugin.Node.Logger.Debugw("DisconnectNodes", "sourceID", sourceId, "targetId", targetId)
 		lock.Lock()
 		defer lock.Unlock()
 
@@ -69,8 +75,30 @@ func Configure(plugin *node.Plugin) {
 	}))
 }
 
-func Replay(handlers *types.EventHandlers) {
+func getEventsToReplay() (map[string]bool, map[string]map[string]bool) {
+	lock.Lock()
+	defer lock.Unlock()
+
+	copiedNodes := make(map[string]bool)
 	for nodeId, online := range nodes {
+		copiedNodes[nodeId] = online
+	}
+
+	copiedLinks := make(map[string]map[string]bool)
+	for sourceId, targetMap := range links {
+		copiedLinks[sourceId] = make(map[string]bool)
+		for targetId := range targetMap {
+			copiedLinks[sourceId][targetId] = true
+		}
+	}
+
+	return copiedNodes, copiedLinks
+}
+
+func Replay(handlers *types.EventHandlers) {
+	copiedNodes, copiedLinks := getEventsToReplay()
+
+	for nodeId, online := range copiedNodes {
 		handlers.AddNode(nodeId)
 		if online {
 			handlers.NodeOnline(nodeId)
@@ -79,7 +107,7 @@ func Replay(handlers *types.EventHandlers) {
 		}
 	}
 
-	for sourceId, targetMap := range links {
+	for sourceId, targetMap := range copiedLinks {
 		for targetId := range targetMap {
 			handlers.ConnectNodes(sourceId, targetId)
 		}
