@@ -1,16 +1,20 @@
 package local
 
 import (
-	"fmt"
-	"io/ioutil"
+	"crypto/ed25519"
+	"encoding/base64"
 	"net"
-	"net/http"
 	"strconv"
+	"strings"
 	"sync"
 
-	"github.com/iotaledger/goshimmer/packages/autopeering/peer"
-	"github.com/iotaledger/goshimmer/packages/parameter"
+	"github.com/iotaledger/hive.go/autopeering/peer"
+	"github.com/iotaledger/hive.go/autopeering/peer/service"
 	"github.com/iotaledger/hive.go/logger"
+	"github.com/iotaledger/hive.go/netutil"
+
+	"github.com/iotaledger/goshimmer/packages/database"
+	"github.com/iotaledger/goshimmer/plugins/config"
 )
 
 var (
@@ -21,63 +25,63 @@ var (
 func configureLocal() *peer.Local {
 	log := logger.NewLogger("Local")
 
-	ip := net.ParseIP(parameter.NodeConfig.GetString(CFG_ADDRESS))
-	if ip == nil {
-		log.Fatalf("Invalid IP address: %s", parameter.NodeConfig.GetString(CFG_ADDRESS))
-	}
-	if ip.IsUnspecified() {
-		log.Info("Querying public IP ...")
-		myIp, err := getPublicIP(isIPv4(ip))
+	var externalIP net.IP
+	if str := config.Node.GetString(CFG_EXTERNAL); strings.ToLower(str) == "auto" {
+		log.Info("Querying external IP ...")
+		ip, err := netutil.GetPublicIP(false)
 		if err != nil {
-			log.Fatalf("Error querying public IP: %s", err)
+			log.Fatalf("Error querying external IP: %s", err)
 		}
-		ip = myIp
-		log.Infof("Public IP queried: address=%s", ip.String())
+		log.Infof("External IP queried: address=%s", ip.String())
+		externalIP = ip
+	} else {
+		externalIP = net.ParseIP(str)
+		if externalIP == nil {
+			log.Fatalf("Invalid IP address (%s): %s", CFG_EXTERNAL, str)
+		}
+	}
+	if !externalIP.IsGlobalUnicast() {
+		log.Warnf("IP is not a global unicast address: %s", externalIP.String())
 	}
 
-	port := strconv.Itoa(parameter.NodeConfig.GetInt(CFG_PORT))
+	peeringPort := strconv.Itoa(config.Node.GetInt(CFG_PORT))
 
-	// create a new local node
-	db := peer.NewPersistentDB(log)
+	// announce the peering service
+	services := service.New()
+	services.Update(service.PeeringKey, "udp", net.JoinHostPort(externalIP.String(), peeringPort))
 
-	local, err := peer.NewLocal("udp", net.JoinHostPort(ip.String(), port), db)
+	// set the private key from the seed provided in the config
+	var seed [][]byte
+	if str := config.Node.GetString(CFG_SEED); str != "" {
+		bytes, err := base64.StdEncoding.DecodeString(str)
+		if err != nil {
+			log.Fatalf("Invalid %s: %s", CFG_SEED, err)
+		}
+		if l := len(bytes); l != ed25519.SeedSize {
+			log.Fatalf("Invalid %s length: %d, need %d", CFG_SEED, l, ed25519.SeedSize)
+		}
+		seed = append(seed, bytes)
+	}
+	badgerDB, err := database.Get(database.DBPrefixAutoPeering, database.GetBadgerInstance())
+	if err != nil {
+		log.Fatalf("Error loading DB: %s", err)
+	}
+	peerDB, err := peer.NewDB(badgerDB)
+	if err != nil {
+		log.Fatalf("Error creating peer DB: %s", err)
+	}
+
+	// the private key seed of the current local can be returned the following way:
+	// key, _ := peerDB.LocalPrivateKey()
+	// fmt.Println(base64.StdEncoding.EncodeToString(ed25519.PrivateKey(key).Seed()))
+
+	local, err := peer.NewLocal(services, peerDB, seed...)
 	if err != nil {
 		log.Fatalf("Error creating local: %s", err)
 	}
 	log.Infof("Initialized local: %v", local)
 
 	return local
-}
-
-func isIPv4(ip net.IP) bool {
-	return ip.To4() != nil
-}
-
-func getPublicIP(ipv4 bool) (net.IP, error) {
-	var url string
-	if ipv4 {
-		url = "https://api.ipify.org"
-	} else {
-		url = "https://api6.ipify.org"
-	}
-	resp, err := http.Get(url)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	// the body only consists of the ip address
-	ip := net.ParseIP(string(body))
-	if ip == nil {
-		return nil, fmt.Errorf("not an IP: %s", body)
-	}
-
-	return ip, nil
 }
 
 func GetInstance() *peer.Local {
