@@ -2,30 +2,20 @@ package cli
 
 import (
 	"fmt"
-	"strings"
+	"os"
 
-	"github.com/iotaledger/goshimmer/packages/parameter"
 	"github.com/iotaledger/hive.go/events"
-	"github.com/iotaledger/hive.go/logger"
 	"github.com/iotaledger/hive.go/node"
 	flag "github.com/spf13/pflag"
+
+	"github.com/iotaledger/goshimmer/plugins/banner"
 )
 
-const (
-	// AppVersion version number
-	AppVersion = "v0.1.0"
-	// AppName app code name
-	AppName = "GoShimmer"
-)
+var PLUGIN = node.NewPlugin("CLI", node.Enabled)
 
-var PLUGIN = node.NewPlugin("CLI", node.Enabled, configure, run)
-
-func onAddPlugin(name string, status int) {
-	AddPluginStatus(node.GetPluginIdentifier(name), status)
-}
+var version = flag.BoolP("version", "v", false, "Prints the GoShimmer version")
 
 func init() {
-
 	for name, status := range node.GetPlugins() {
 		onAddPlugin(name, status)
 	}
@@ -33,44 +23,17 @@ func init() {
 	node.Events.AddPlugin.Attach(events.NewClosure(onAddPlugin))
 
 	flag.Usage = printUsage
+
+	PLUGIN.Events.Init.Attach(events.NewClosure(onInit))
 }
 
-func parseParameters() {
-	for _, pluginName := range parameter.NodeConfig.GetStringSlice(node.CFG_DISABLE_PLUGINS) {
-		node.DisabledPlugins[strings.ToLower(pluginName)] = true
-	}
-	for _, pluginName := range parameter.NodeConfig.GetStringSlice(node.CFG_ENABLE_PLUGINS) {
-		node.EnabledPlugins[strings.ToLower(pluginName)] = true
-	}
+func onAddPlugin(name string, status int) {
+	AddPluginStatus(node.GetPluginIdentifier(name), status)
 }
 
-func LoadConfig() {
-	if err := parameter.FetchConfig(false); err != nil {
-		panic(err)
+func onInit(*node.Plugin) {
+	if *version {
+		fmt.Println(banner.AppName + " " + banner.AppVersion)
+		os.Exit(0)
 	}
-	parseParameters()
-
-	if err := logger.InitGlobalLogger(parameter.NodeConfig); err != nil {
-		panic(err)
-	}
-}
-
-func configure(ctx *node.Plugin) {
-	fmt.Printf(`
-   _____  ____   _____ _    _ _____ __  __ __  __ ______ _____  
-  / ____|/ __ \ / ____| |  | |_   _|  \/  |  \/  |  ____|  __ \ 
- | |  __| |  | | (___ | |__| | | | | \  / | \  / | |__  | |__) |
- | | |_ | |  | |\___ \|  __  | | | | |\/| | |\/| |  __| |  _  / 
- | |__| | |__| |____) | |  | |_| |_| |  | | |  | | |____| | \ \ 
-  \_____|\____/|_____/|_|  |_|_____|_|  |_|_|  |_|______|_|  \_\
-                             %s                                     
-`, AppVersion)
-	fmt.Println()
-
-	ctx.Node.Logger.Infof("GoShimmer version %s ...", AppVersion)
-	ctx.Node.Logger.Info("Loading plugins ...")
-}
-
-func run(ctx *node.Plugin) {
-	// do nothing; everything is handled in the configure step
 }
