@@ -1,26 +1,22 @@
 package graph
 
 import (
-	"github.com/iotaledger/goshimmer/packages/parameter"
+	"github.com/iotaledger/goshimmer/plugins/banner"
+	flag "github.com/spf13/pflag"
+)
+
+const (
+	CFG_WEBROOT      = "graph.webrootPath"
+	CFG_SOCKET_IO    = "graph.socketioPath"
+	CFG_DOMAIN       = "graph.domain"
+	CFG_BIND_ADDRESS = "graph.bindAddress"
+	CFG_NETWORK      = "graph.networkName"
 )
 
 func init() {
-
-	// "Path to IOTA Tangle Visualiser webroot files"
-	parameter.NodeConfig.SetDefault("graph.webrootPath", "IOTAtangle/webroot")
-
-	// "Path to socket.io.js"
-	parameter.NodeConfig.SetDefault("graph.socketioPath", "socket.io-client/dist/socket.io.js")
-
-	// "Set the domain on which IOTA Tangle Visualiser is served"
-	parameter.NodeConfig.SetDefault("graph.domain", "")
-
-	// "Set the host to which the IOTA Tangle Visualiser listens"
-	parameter.NodeConfig.SetDefault("graph.host", "127.0.0.1")
-
-	// "IOTA Tangle Visualiser webserver port"
-	parameter.NodeConfig.SetDefault("graph.port", 8083)
-
-	// "Name of the network shown in IOTA Tangle Visualiser"
-	parameter.NodeConfig.SetDefault("graph.networkName", "meets HORNET")
+	flag.String(CFG_WEBROOT, "IOTAtangle/webroot", "Path to IOTA Tangle Visualiser webroot files")
+	flag.String(CFG_SOCKET_IO, "socket.io-client/dist/socket.io.js", "Path to socket.io.js")
+	flag.String(CFG_DOMAIN, "", "Set the domain on which IOTA Tangle Visualiser is served")
+	flag.String(CFG_BIND_ADDRESS, "127.0.0.1:8082", "the bind address for the IOTA Tangle Visualizer")
+	flag.String(CFG_NETWORK, banner.AppName, "Name of the network shown in IOTA Tangle Visualiser")
 }
