@@ -4,10 +4,10 @@ import (
 	"encoding/base64"
 	"net/http"
 
-	"github.com/iotaledger/goshimmer/packages/autopeering/peer"
-	"github.com/iotaledger/goshimmer/packages/autopeering/peer/service"
 	"github.com/iotaledger/goshimmer/plugins/autopeering"
 	"github.com/iotaledger/goshimmer/plugins/webapi"
+	"github.com/iotaledger/hive.go/autopeering/peer"
+	"github.com/iotaledger/hive.go/autopeering/peer/service"
 	"github.com/iotaledger/hive.go/node"
 	"github.com/labstack/echo"
 )
@@ -23,13 +23,28 @@ func getNeighbors(c echo.Context) error {
 
 	chosen := []Neighbor{}
 	accepted := []Neighbor{}
+	knownPeers := []Neighbor{}
 
 	if autopeering.Selection == nil {
-		return requestFailed(c, "Neighbor Selection is not enabled")
+		return c.JSON(http.StatusNotImplemented, Response{Error: "Neighbor Selection is not enabled"})
+	}
+
+	if autopeering.Discovery == nil {
+		return c.JSON(http.StatusNotImplemented, Response{Error: "Neighbor Discovery is not enabled"})
+	}
+
+	if c.QueryParam("known") == "1" {
+		for _, peer := range autopeering.Discovery.GetVerifiedPeers() {
+			n := Neighbor{
+				ID:        peer.ID().String(),
+				PublicKey: base64.StdEncoding.EncodeToString(peer.PublicKey()),
+			}
+			n.Services = getServices(peer)
+			knownPeers = append(knownPeers, n)
+		}
 	}
 
 	for _, peer := range autopeering.Selection.GetOutgoingNeighbors() {
-
 		n := Neighbor{
 			ID:        peer.ID().String(),
 			PublicKey: base64.StdEncoding.EncodeToString(peer.PublicKey()),
@@ -45,33 +60,21 @@ func getNeighbors(c echo.Context) error {
 		n.Services = getServices(peer)
 		accepted = append(accepted, n)
 	}
-	return requestSuccessful(c, chosen, accepted)
 
-}
-
-func requestSuccessful(c echo.Context, chosen, accepted []Neighbor) error {
-	return c.JSON(http.StatusOK, Response{
-		Chosen:   chosen,
-		Accepted: accepted,
-	})
-}
-
-func requestFailed(c echo.Context, message string) error {
-	return c.JSON(http.StatusNotFound, Response{
-		Error: message,
-	})
+	return c.JSON(http.StatusOK, Response{KnownPeers: knownPeers, Chosen: chosen, Accepted: accepted})
 }
 
 type Response struct {
-	Chosen   []Neighbor `json:"chosen"`
-	Accepted []Neighbor `json:"accepted"`
-	Error    string     `json:"error,omitempty"`
+	KnownPeers []Neighbor `json:"known,omitempty"`
+	Chosen     []Neighbor `json:"chosen"`
+	Accepted   []Neighbor `json:"accepted"`
+	Error      string     `json:"error,omitempty"`
 }
 
 type Neighbor struct {
-	ID        string `json:"id"`        // comparable node identifier
-	PublicKey string `json:"publicKey"` // public key used to verify signatures
-	Services  []peerService
+	ID        string        `json:"id"`        // comparable node identifier
+	PublicKey string        `json:"publicKey"` // public key used to verify signatures
+	Services  []peerService `json:"services,omitempty"`
 }
 
 type peerService struct {
