@@ -3,10 +3,10 @@ package gtta
 import (
 	"net/http"
 
-	"github.com/iotaledger/goshimmer/plugins/tipselection"
+	"github.com/iotaledger/goshimmer/packages/binary/tangle/model/transaction"
+	"github.com/iotaledger/goshimmer/plugins/tangle"
 	"github.com/iotaledger/goshimmer/plugins/webapi"
 	"github.com/iotaledger/hive.go/node"
-	"github.com/iotaledger/iota.go/trinary"
 	"github.com/labstack/echo"
 )
 
@@ -15,17 +15,15 @@ var PLUGIN = node.NewPlugin("WebAPI GTTA Endpoint", node.Disabled, func(plugin *
 })
 
 func Handler(c echo.Context) error {
-
-	branchTransactionHash := tipselection.GetRandomTip()
-	trunkTransactionHash := tipselection.GetRandomTip()
+	trunkTransactionId, branchTransactionId := tangle.TipSelector.GetTips()
 
 	return c.JSON(http.StatusOK, Response{
-		BranchTransaction: branchTransactionHash,
-		TrunkTransaction:  trunkTransactionHash,
+		TrunkTransaction:  trunkTransactionId,
+		BranchTransaction: branchTransactionId,
 	})
 }
 
 type Response struct {
-	BranchTransaction trinary.Trytes `json:"branchTransaction"`
-	TrunkTransaction  trinary.Trytes `json:"trunkTransaction"`
+	BranchTransaction transaction.Id `json:"branchTransaction"`
+	TrunkTransaction  transaction.Id `json:"trunkTransaction"`
 }
